+++ packages/rich-text/src/plugins/Hr/index.tsx
@@ -39,7 +39,7 @@ const styles = {
       position: absolute;
       width: 100%;
       height: 1px;
-      background: ${tokens.colorElementMid};
+      background: ${tokens.gray300};
       top: 50%;
     }
   `,