+++ packages/location/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.12.1...@contentful/field-editor-location@0.12.2) (2021-08-19)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+## [0.12.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.12.0...@contentful/field-editor-location@0.12.1) (2021-07-29)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+# [0.12.0](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.6...@contentful/field-editor-location@0.12.0) (2021-07-23)
+
+
+### Features
+
+* 💡 new color tokens ([#778](https://github.com/contentful/field-editors/issues/778)) ([fba548d](https://github.com/contentful/field-editors/commit/fba548de32305016df7f2685634eefb14294828f))
+
+
+
+
+
+## [0.11.6](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.3...@contentful/field-editor-location@0.11.6) (2021-07-06)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+## [0.11.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.3...@contentful/field-editor-location@0.11.5) (2021-07-06)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+## [0.11.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.3...@contentful/field-editor-location@0.11.4) (2021-06-23)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+## [0.11.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.2...@contentful/field-editor-location@0.11.3) (2021-06-23)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
+## [0.11.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.1...@contentful/field-editor-location@0.11.2) (2021-06-22)
+
+**Note:** Version bump only for package @contentful/field-editor-location
+
+
+
+
+
 ## [0.11.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-location@0.11.0...@contentful/field-editor-location@0.11.1) (2021-03-05)
 
 **Note:** Version bump only for package @contentful/field-editor-location